refactor: split redux hooks
