+++ app/src/acarsSlice.ts
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit'
 import type { PayloadAction } from '@reduxjs/toolkit'
-import { useAppDispatch, useAppSelector } from './store'
+import { useAppDispatch, useAppSelector } from './storeHooks'
 import type { AppDispatch, RootState } from './store'
 
 export interface DataPoint {
